Split gulp file
