+++ package.json
@@ -6,20 +6,15 @@
   "license": "Apache-2.0",
   "author": "Google",
   "repository": "google/material-design-lite",
-  "devDependencies": {
-    "babel-core": "^5.8.25",
-    "gulp": "gulpjs/gulp#4.0"
-  },
   "engines": {
     "node": ">=0.12.0"
   },
   "scripts": {
     "test": "gulp && git status | grep 'working directory clean' >/dev/null || (echo 'Please commit all changes generated by building'; exit 1)"
   },
-  "babel": {
-    "only": "gulpfile.babel.js"
-  },
   "dependencies": {
+    "babel-core": "^5.8.25",
+    "gulp": "gulpjs/gulp#4.0",
     "acorn": "^2.6.4",
     "browser-sync": "^2.10.0",
     "chai": "^3.4.1",
