+++ utils/customizer.js
@@ -95,7 +95,8 @@ function init() {
   req.send();
 }
 
-MaterialCustomizer = (function() {
+// Global!
+var MaterialCustomizer = (function() {
   'use strict';
 
   var COLORS = ['Cyan', 'Teal', 'Green', 'Light Green', 'Lime',
